+++ test/tests.py
@@ -1,10 +1,24 @@
 import os
 import sys
-sys.path.insert(0, os.getcwd())
-from lca_algebraic import *
 import pytest
 
-initDb('MyProject')
+sys.path.insert(0, os.getcwd())
+sys.path.insert(0, os.path.join(os.getcwd(), "test"))
+
+from lca_algebraic import *
+from fixtures import *
+
+USER_DB = "fg"
+BG_DB= "bg"
+METHOD_PREFIX='tests'
+
+SET_USER_DB(USER_DB)
+SET_BG_DB(BG_DB)
+
+# Reset func project, empty DB
+initDb('tests')
+init_acts(BG_DB)
+init_methods(BG_DB, METHOD_PREFIX)
 
 # Import Ecoinvent DB (if not already done)
 # Update the name and path to the location of the ecoinvent database
@@ -12,13 +26,13 @@ initDb('MyProject')
 
 # We use a separate DB for defining our foreground model / activities
 # Choose any name
-USER_DB = 'test-db'
+
 
 def setup_function() :
     """Before each test"""
 
-    print("resetting DB")
-    SET_USER_DB(USER_DB)
+    print("resetting fg DB")
+
     resetDb(USER_DB)
     resetParams(USER_DB)
 
@@ -32,22 +46,39 @@ def test_switch_activity_support_sevral_times_same_target() :
         values=["v1", "v2", "v3"],
         default="v1")
 
-    alu = findTechAct("aluminium alloy production, AlMg3", loc="RER")
+    bg_act1 = findTechAct("bg_act1")
 
     act = newSwitchAct(USER_DB, "switchAct", p1, {
-        "v1" : alu,
-        "v2" : alu,
-        "v3" : alu
+        "v1" : bg_act1,
+        "v2" : bg_act1,
+        "v3" : bg_act1
     })
 
-    climate = [m for m in bw.methods if 'ILCD 1.0.8 2016' in str(m) and 'no LT' in str(m)][1]
+    impact = (METHOD_PREFIX, 'all', 'total')
 
-    res = multiLCAAlgebric(act, [climate], p1=["v1", "v2", "v3"])
+    res = multiLCAAlgebric(act, [impact], p1=["v1", "v2", "v3"])
     vals = res.values
 
     assert vals[0] == vals[1] and vals[1] == vals[2]
 
-def test_enum_vales_are_enforced():
+def test_new_switch_act_with_tuples() :
+
+    p1 = newEnumParam(
+        'p1',
+        values=["v1", "v2", "v3"],
+        default="v1")
+
+    bg_act1 = findTechAct("bg_act1")
+    bg_act2 = findTechAct("bg_act2")
+
+    newSwitchAct(USER_DB, "switchAct", p1, {
+        "v1": (bg_act1, 2.0),
+        "v2": (bg_act2, 3.0),
+    })
+
+
+
+def test_enum_values_are_enforced():
 
     # Enum param
     p1 = newEnumParam(
@@ -60,7 +91,6 @@ def test_enum_vales_are_enforced():
 
     with pytest.raises(Exception) as exc:
         multiLCAAlgebric(act, climate, p1="bar")
-
 
     assert 'Invalid value' in str(exc)